Move folder-selector code to a separate helper function so we can de-duplicate it
